Calculate the average of a list of numbers
The objective of this task is to create a function that takes an array of numbers as input and calculates the average of those numbers. An average is a measure of central tendency, indicating the 'middle' value of a set of numbers. In this case, the average will be calculated by dividing the sum of all the numbers in the array by the number of numbers in the array